Se cambio de ubicacion el schema de pedido, era una ubicacion erronea, y se agrego el un array para almacenar los productos de los pedidos.)
