+++ src/database/model/pedido.js
@@ -1,0 +1,28 @@
+import mongoose, { Schema } from 'mongoose';
+
+const pedidoSchema = new Schema({
+  usuario: {
+    type: String,
+    required: true,
+    minLength: 8,
+    maxLength: 50,
+  },
+  monto: {
+    type: Number,
+    required: true,
+  },
+  productos: [{ type: Schema.Types.ObjectId, ref: 'producto' }],
+  metodoEnvio: {
+    type: Number,
+    required: true,
+    enum: [1, 2],
+  },
+  estadoEnvio: {
+    type: Boolean,
+    required: true,
+  },
+});
+
+const Pedido = mongoose.model('pedido', pedidoSchema);
+
+export default Pedido;
